admin/requirements.txt
Flask==3.0.0
Flask-WTF
google-cloud-datastore==2.15.1
google-auth
requests==2.28.2
gunicorn
